MMR Search for Pinecone, Weaviate, Zep and Supabase
